Folder restructuring and charts refactoring.
