+++ src/shared/components/cards/InfoCard.tsx
